cleanup the api tests
