+++ e2e/globalSetup.ts
@@ -3,16 +3,16 @@ import { chromium, FullConfig } from '@playwright/test';
 import { IDENTIFIER, TOKEN } from './nextAuth.constants';
 
 async function globalSetup(config: FullConfig) {
-  const { baseURL, storageState } = config.projects[0].use;
-  // Generate a link with email, unhashed token and callback url
-  const params = new URLSearchParams({ callbackURL: baseURL || '', token: TOKEN, email: IDENTIFIER });
-  const providerId = 'email';
-  const MAGIC_LINK = `${baseURL}/api/auth/callback/${providerId}?${params}`;
-  const browser = await chromium.launch();
-  const page = await browser.newPage();
-  await page.goto(MAGIC_LINK);
-  await page.context().storageState({ path: storageState as string });
-  await browser.close();
+  // const { baseURL, storageState } = config.projects[0].use;
+  // // Generate a link with email, unhashed token and callback url
+  // const params = new URLSearchParams({ callbackURL: baseURL || '', token: TOKEN, email: IDENTIFIER });
+  // const providerId = 'email';
+  // const MAGIC_LINK = `${baseURL}/api/auth/callback/${providerId}?${params}`;
+  // const browser = await chromium.launch();
+  // const page = await browser.newPage();
+  // await page.goto(MAGIC_LINK);
+  // await page.context().storageState({ path: storageState as string });
+  // await browser.close();
 }
 
 export default globalSetup;
